Cleaned up DB connection code
